// packages/stitch/src/protoc/output/msp/msp_principal_pb.d.ts
export class MSPPrincipal extends jspb.Message {
  getPrincipalClassification(): MSPPrincipal.Classification;
  setPrincipalClassification(value: MSPPrincipal.Classification): void;

  getPrincipal(): Uint8Array | string;
  getPrincipal_asU8(): Uint8Array;
  getPrincipal_asB64(): string;
  setPrincipal(value: Uint8Array | string): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): MSPPrincipal.AsObject;
  static toObject(includeInstance: boolean, msg: MSPPrincipal): MSPPrincipal.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: MSPPrincipal, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): MSPPrincipal;
  static deserializeBinaryFromReader(message: MSPPrincipal, reader: jspb.BinaryReader): MSPPrincipal;
}

export namespace MSPPrincipal {
  export type AsObject = {
    principalClassification: MSPPrincipal.Classification,
    principal: Uint8Array | string,
  }

  export enum Classification {
    ROLE = 0,
    ORGANIZATION_UNIT = 1,
    IDENTITY = 2,
    ANONYMITY = 3,
    COMBINED = 4,
  }
}
// ...
export class MSPRole extends jspb.Message {
  getMspIdentifier(): string;
  setMspIdentifier(value: string): void;

  getRole(): MSPRole.MSPRoleType;
  setRole(value: MSPRole.MSPRoleType): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): MSPRole.AsObject;
  static toObject(includeInstance: boolean, msg: MSPRole): MSPRole.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: MSPRole, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): MSPRole;
  static deserializeBinaryFromReader(message: MSPRole, reader: jspb.BinaryReader): MSPRole;
}

export namespace MSPRole {
  export type AsObject = {
    mspIdentifier: string,
    role: MSPRole.MSPRoleType,
  }

  export enum MSPRoleType {
    MEMBER = 0,
    ADMIN = 1,
    CLIENT = 2,
    PEER = 3,
  }
}